Reorganize the Linux implementation and split OSMesa and X11
