+++ src/api/mod.rs
@@ -1,5 +1,6 @@
 pub mod android;
 pub mod cocoa;
 pub mod egl;
+pub mod osmesa;
 pub mod win32;
 pub mod x11;
