+++ src/api/x11/monitor.rs
@@ -1,6 +1,6 @@
 use std::ptr;
 use std::collections::VecDeque;
-use super::super::ffi;
+use super::ffi;
 use super::ensure_thread_init;
 use native_monitor::NativeMonitorId;
